Move DecaySpectrum to distribution.h/cpp
